Move coordinate system code to common submodule
